update-a-07-11-2021

+++ CvTemplate Solution/CvTemplate.Domain/Models/Entities/BlogPostComment.cs
@@ -10,9 +10,9 @@ namespace CvTemplate.Domain.Models.Entities
     public class BlogPostComment : BaseEntity
     {
         public string Comment { get; set; }
-        public long BlogPostId { get; set; }
+        public int BlogPostId { get; set; }
         public virtual BlogPost BlogPost { get; set; }
-        public long? ParentId { get; set; }
+        public int? ParentId { get; set; }
         public virtual BlogPostComment Parent { get; set; }
         public virtual ICollection<BlogPostComment> Children { get; set; }
         public virtual CvTemplateUser CreatedByUser { get; set; }
